Unify regression lines and branch display across layouts
The clock view is simply a specific type of scatterplot layout. This commit adds relevant toggles to both the clock layout selector and the scatterplot layout selector.

To improve the UX, once a scatterplot has been viewed, we persist the x,y variables for future viewing. Similarly, the toggle state persists between clock & scatter layouts.

This commit also makes regression lines available for both layouts. Previously, the regression would be shown for clock layouts _if_ the branch metric was time. Here we expose this option as a UI toggle, with he regression calculation unchanged. For scatter-plot views we calculate the regression with a free intercept, as the root node may not have co-ordinates defined (depending on chosen x,y variables), and additionally report the R^2. The display of the regression text can be improved in future commits.